backend/legislative_analysis_func/requirements.txt
functions-framework==3.*
google-cloud-aiplatform==1.56.0
google-cloud-storage==2.*
google-cloud-firestore==2.*
google-auth==2.23.4